local remote mediator cleanup
